Update to discord.js v14
This is a first-pass attempt at updating dependencies with no other changes.
discord.js v13 & v14 include a new command mechanism (slash commands) and a
fork/rework of the voice package in ways that require a more thorough
rearchitecting of this bot. I don't running the bot from this commit onward
until that rework has occurred.